[ADD] ticoelectronics: Purchase Price Over Limit Warning
Adding a new field in the models `res.company` and `res.config.settings`
called `allowed_percentage_unit_price` that will be sed to allow
exceeding the last unit price used for every product in all the purchase
orders.

Opening a wizard with the warning when at least one of the lines in a
`purchase.order` exceeds the percentage in the
`allowed_percentage_unit_price` field or more from the last unit price
(converted to the price in the order), where they will be able to decide
if they want to confirm the purchase anyways or if they don't want to
confirm it.

Also, setting the `allowed_percentage_unit_price` of the main company at
5%.